pyproject.toml
[project]
name = "automated-task-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.56.0",
    "chatlas>=0.9.0",
    "crawl4ai>=0.6.3",
    "faiss-cpu>=1.11.0",
    "fastmcp>=2.10.1",
    "google-genai>=1.24.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.16",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.6",
    "langchain-openai>=0.3.27",
    "langgraph>=0.2.31",
    "mcp[cli]>=1.10.1",
    "networkx>=3.5",
    "openai>=1.93.0",
    "pydantic>=2.11.7",
    "pydantic-ai>=0.3.5",
    "pyrsm>=1.6.0",
    "python-dotenv>=1.1.1",
    "pyvis>=0.3.2",
    "requests>=2.32.4",
    "sentence-transformers>=5.0.0",
    "smolagents[mcp]>=1.19.0",
    "streamlit-calendar>=1.3.1",
]
